Add the get powers by hero endpoint
