+++ server/src/queries.js
@@ -4,7 +4,8 @@ export const INSERT_HERO = 'INSERT INTO heroes SET ?';
 export const UPDATE_HERO = 'UPDATE heroes SET ? WHERE id = ?';
 export const DELETE_HERO = 'DELETE FROM heroes WHERE id = ?';
 export const SELECT_VEHICLES = 'SELECT * FROM vehicles';
-export const SELECT_VEHICLES_BY_HEROE = 'SELECT * FROM vehicles WHERE id = ?';
+export const SELECT_VEHICLES_BY_HEROE = 'SELECT * FROM vehicles WHERE heroes_id = ?';
+export const SELECT_POWERS_BY_HEROE = 'SELECT heroes_has_powers.heroes_id, heroes_has_powers.powers_id, heroes.name as hero, powers.name as power FROM heroes_has_powers, heroes, powers WHERE heroes_has_powers.heroes_id = heroes.id AND heroes_has_powers.powers_id = powers.id AND heroes_has_powers.heroes_id = ?';
 export const SELECT_VEHICLE = 'SELECT * FROM vehicles WHERE id = ?';
 export const INSERT_VEHICLE = 'INSERT INTO vehicles SET ?';
 export const UPDATE_VEHICLE = 'UPDATE vehicles SET ? WHERE id = ?';
